Make relationfield serializer/deserializer optional
