+++ src/plone/restapi/serializer/configure.zcml
@@ -38,7 +38,6 @@
     <adapter factory=".converters.persistent_list_converter" />
     <adapter factory=".converters.persistent_mapping_converter" />
     <adapter factory=".converters.python_datetime_converter" />
-    <adapter factory=".converters.relationvalue_converter" />
     <adapter factory=".converters.richtext_converter" />
     <adapter factory=".converters.set_converter" />
     <adapter factory=".converters.string_converter" />
@@ -46,5 +45,9 @@
     <adapter factory=".converters.timedelta_converter" />
     <adapter factory=".converters.tuple_converter" />
     <adapter factory=".converters.zope_DateTime_converter" />
+
+    <configure zcml:condition="installed z3c.relationfield">
+        <adapter factory=".relationfield.relationvalue_converter" />
+    </configure>
 
 </configure>
