+++ src/plone/restapi/serializer/converters.py
@@ -10,13 +10,9 @@ from persistent.list import PersistentList
 from persistent.mapping import PersistentMapping
 from plone.app.textfield.interfaces import IRichTextValue
 from plone.restapi.interfaces import IJsonCompatible
-from plone.restapi.interfaces import ISerializeToJsonSummary
-from z3c.relationfield.interfaces import IRelationValue
 from zope.component import adapter
-from zope.component import getMultiAdapter
 from zope.component.hooks import getSite
 from zope.interface import Interface
-from zope.globalrequest import getRequest
 from zope.interface import implementer
 
 
@@ -139,12 +135,3 @@ def richtext_converter(value):
         u'content-type': json_compatible(value.mimeType),
         u'encoding': json_compatible(value.encoding),
     }
-
-
-@adapter(IRelationValue)
-@implementer(IJsonCompatible)
-def relationvalue_converter(value):
-    if value.to_object:
-        summary = getMultiAdapter(
-            (value.to_object, getRequest()), ISerializeToJsonSummary)()
-        return json_compatible(summary)
